refactor: 좌표값 기존 int -> long 변경
- 상대 좌표 %가 소수점 가

+++ src/main/java/org/complete/challang/app/combination/controller/dto/item/CombinationFindDto.java
@@ -20,10 +20,10 @@ public class CombinationFindDto {
     private String volume;
 
     @JsonProperty("xcoordinate")
-    private int xCoordinate;
+    private float xCoordinate;
 
     @JsonProperty("ycoordinate")
-    private int yCoordinate;
+    private float yCoordinate;
 
     public static CombinationFindDto toDto(Combination combination,
                                            Long userId) {
